Refactor provider panel rendering
